python-services/screening-worker/requirements.txt
kafka-python>=2.0.0
pandas>=2.2.0
psycopg2-binary>=2.9.9
